test params parsing
