+++ reduct_cli/Cargo.toml
@@ -24,6 +24,7 @@ chrono = "0.4.34"
 dialoguer = "0.11.0"
 futures-util = { version = "0.3.30", features = [] }
 indicatif= "0.17.8"
+async-trait = "0.1.78"
 
 [dev-dependencies]
 rstest = "0.18.2"
