+++ reduct_cli/src/cmd/cp.rs
@@ -3,14 +3,15 @@
 //    License, v. 2.0. If a copy of the MPL was not distributed with this
 //    file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-mod bucket_to_bucket;
+mod b2b;
+mod helpers;
 
-use crate::cmd::cp::bucket_to_bucket::cp_bucket_to_bucket;
+use crate::cmd::cp::b2b::cp_bucket_to_bucket;
 use crate::cmd::ALIAS_OR_URL_HELP;
 use crate::context::CliContext;
 use crate::parsers::ResourcePathParser;
 use clap::ArgAction::{SetFalse, SetTrue};
-use clap::{Arg, Command};
+use clap::{value_parser, Arg, Command};
 use url::Url;
 
 pub(crate) fn cp_cmd() -> Command {
@@ -47,6 +48,7 @@ pub(crate) fn cp_cmd() -> Command {
             Arg::new("include")
                 .long("include")
                 .short('I')
+                .value_name("KEY=VALUE")
                 .help("Only these records which have this key-value pair will be exported.\nThe format is key=value. This option can be used multiple times to include multiple key-value pairs.")
                 .num_args(1..)
                 .required(false)
@@ -55,6 +57,7 @@ pub(crate) fn cp_cmd() -> Command {
             Arg::new("exclude")
                 .long("exclude")
                 .short('E')
+                .value_name("KEY=VALUE")
                 .help("These records which have this key-value pair will not be exported.\nThe format is key=value. This option can be used multiple times to exclude multiple key-value pairs.")
                 .num_args(1..)
                 .required(false)
@@ -62,6 +65,7 @@ pub(crate) fn cp_cmd() -> Command {
         .arg(Arg::new("entry")
             .long("entry")
             .short('n')
+            .value_name("ENTRY_NAME")
             .help("List of entries to export.\nIf not specified, all entries will be exported. Wildcards are supported.")
             .num_args(1..)
             .required(false)
@@ -72,6 +76,7 @@ pub(crate) fn cp_cmd() -> Command {
                 .short('l')
                 .help("The maximum number of records to export.\nIf not specified, all recordswill be exported.")
                 .value_name("NUMBER")
+                .value_parser(value_parser!(u64))
                 .required(false)
         )
         .arg(
